refactor(task): extract stage tests into graphresolver_stage_test.go and co-locate stageTask
