+++ pkg/core/task/graphresolver_cycle_test.go
@@ -435,259 +435,3 @@ func TestResolveGraph_ShuffleInvariance(t *testing.T) {
 		}
 	}
 }
-
-func TestResolveGraph_FanInCycle_DownstreamTaskPtPRouting(t *testing.T) {
-	tagA := NewTag[any]("tag-a")
-
-	testCases := []struct {
-		name                 string
-		initialTasks         []UntypedTask
-		availableTasks       []UntypedTask
-		wantTaskIDs          []string
-		wantDownstreamSource string
-	}{
-		// Mermaid task graph:
-		// ```mermaid
-		// graph TD
-		//   prod_high["prod-high (priority=10)"] -->|"FanIn (tag-a)"| consumer_s1["consumer#default-stage-1"]
-		//   consumer_s1 -->|"PointToPoint"| prod_low["prod-low (priority=100)"]
-		//   prod_low -->|"FanIn (tag-a)"| consumer_s2["consumer#default-stage-2"]
-		//   prod_high -->|"FanIn (tag-a)"| consumer_s2
-		//   consumer_s1 -.->|"OrderOnly"| consumer_s2
-		//   consumer_s2 -->|"PointToPoint"| downstream["downstream"]
-		// ```
-		{
-			name: "downstream task depending on multi-stage consumer is routed from stage-2 when not leading to feedback",
-			initialTasks: []UntypedTask{
-				createMockTask("prod-high", "default", nil, ProvidesTag(tagA, WithTagPriority(10))),
-				createMockTask("consumer", "default", []Dependency{
-					tagA.Ref(),
-				}, AllowMultiStageExecution()),
-				createMockTask("prod-low", "default", []Dependency{
-					taskid.NewTaskReference[any]("consumer"),
-				}, ProvidesTag(tagA, WithTagPriority(100))),
-				createMockTask("downstream", "default", []Dependency{
-					taskid.NewTaskReference[any]("consumer"),
-				}),
-			},
-			availableTasks: []UntypedTask{
-				createMockTask("prod-high", "default", nil, ProvidesTag(tagA, WithTagPriority(10))),
-				createMockTask("consumer", "default", []Dependency{
-					tagA.Ref(),
-				}, AllowMultiStageExecution()),
-				createMockTask("prod-low", "default", []Dependency{
-					taskid.NewTaskReference[any]("consumer"),
-				}, ProvidesTag(tagA, WithTagPriority(100))),
-				createMockTask("downstream", "default", []Dependency{
-					taskid.NewTaskReference[any]("consumer"),
-				}),
-			},
-			wantTaskIDs: []string{
-				"prod-high#default",
-				"consumer#default-stage-1",
-				"prod-low#default",
-				"consumer#default-stage-2",
-				"downstream#default",
-			},
-			wantDownstreamSource: "consumer#default-stage-2",
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			taskSet, err := ResolveGraph(tc.initialTasks, tc.availableTasks, nil)
-			if err != nil {
-				t.Fatalf("ResolveGraph() error = %v, want nil", err)
-			}
-
-			var gotTaskIDs []string
-			for _, task := range taskSet.GetAll() {
-				gotTaskIDs = append(gotTaskIDs, task.UntypedID().String())
-			}
-			if diff := cmp.Diff(tc.wantTaskIDs, gotTaskIDs); diff != "" {
-				t.Errorf("task order mismatch (-want +got):\n%s", diff)
-			}
-
-			incoming := taskSet.IncomingEdges("downstream#default")
-			if len(incoming) != 1 {
-				t.Fatalf("downstream incoming edges count = %d, want 1", len(incoming))
-			}
-			if gotSource := incoming[0].SourceImplID; gotSource != tc.wantDownstreamSource {
-				t.Errorf("downstream incoming edge source = %q, want %q", gotSource, tc.wantDownstreamSource)
-			}
-		})
-	}
-}
-
-func TestResolveGraph_FanInCycle_UpstreamTaskPtPRouting(t *testing.T) {
-	tagA := NewTag[any]("tag-a")
-
-	testCases := []struct {
-		name               string
-		initialTasks       []UntypedTask
-		availableTasks     []UntypedTask
-		wantTaskIDs        []string
-		wantUpstreamSource string
-	}{
-		// Mermaid task graph:
-		// ```mermaid
-		// graph TD
-		//   upstream["upstream"] -->|"PointToPoint"| consumer_s1["consumer#default-stage-1"]
-		//   upstream -->|"PointToPoint"| consumer_s2["consumer#default-stage-2"]
-		//   prod_high["prod-high (priority=10)"] -->|"FanIn (tag-a)"| consumer_s1
-		//   consumer_s1 -->|"PointToPoint"| prod_low["prod-low (priority=100)"]
-		//   prod_low -->|"FanIn (tag-a)"| consumer_s2
-		//   prod_high -->|"FanIn (tag-a)"| consumer_s2
-		//   consumer_s1 -.->|"OrderOnly"| consumer_s2
-		// ```
-		{
-			name: "upstream task required by multi-stage consumer is routed to both stage-1 and stage-2",
-			initialTasks: []UntypedTask{
-				createMockTask("upstream", "default", nil),
-				createMockTask("prod-high", "default", nil, ProvidesTag(tagA, WithTagPriority(10))),
-				createMockTask("consumer", "default", []Dependency{
-					taskid.NewTaskReference[any]("upstream"),
-					tagA.Ref(),
-				}, AllowMultiStageExecution()),
-				createMockTask("prod-low", "default", []Dependency{
-					taskid.NewTaskReference[any]("consumer"),
-				}, ProvidesTag(tagA, WithTagPriority(100))),
-			},
-			availableTasks: []UntypedTask{
-				createMockTask("upstream", "default", nil),
-				createMockTask("prod-high", "default", nil, ProvidesTag(tagA, WithTagPriority(10))),
-				createMockTask("consumer", "default", []Dependency{
-					taskid.NewTaskReference[any]("upstream"),
-					tagA.Ref(),
-				}, AllowMultiStageExecution()),
-				createMockTask("prod-low", "default", []Dependency{
-					taskid.NewTaskReference[any]("consumer"),
-				}, ProvidesTag(tagA, WithTagPriority(100))),
-			},
-			wantTaskIDs: []string{
-				"prod-high#default",
-				"upstream#default",
-				"consumer#default-stage-1",
-				"prod-low#default",
-				"consumer#default-stage-2",
-			},
-			wantUpstreamSource: "upstream#default",
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			taskSet, err := ResolveGraph(tc.initialTasks, tc.availableTasks, nil)
-			if err != nil {
-				t.Fatalf("ResolveGraph() error = %v, want nil", err)
-			}
-
-			var gotTaskIDs []string
-			for _, task := range taskSet.GetAll() {
-				gotTaskIDs = append(gotTaskIDs, task.UntypedID().String())
-			}
-			if diff := cmp.Diff(tc.wantTaskIDs, gotTaskIDs); diff != "" {
-				t.Errorf("task order mismatch (-want +got):\n%s", diff)
-			}
-
-			incomingS1 := taskSet.IncomingEdges("consumer#default-stage-1")
-			hasUpstreamS1 := false
-			for _, e := range incomingS1 {
-				if e.SourceImplID == tc.wantUpstreamSource {
-					hasUpstreamS1 = true
-					break
-				}
-			}
-			if !hasUpstreamS1 {
-				t.Errorf("expected incoming edge from %q to stage-1, but not found", tc.wantUpstreamSource)
-			}
-
-			incomingS2 := taskSet.IncomingEdges("consumer#default-stage-2")
-			hasUpstreamS2 := false
-			for _, e := range incomingS2 {
-				if e.SourceImplID == tc.wantUpstreamSource {
-					hasUpstreamS2 = true
-					break
-				}
-			}
-			if !hasUpstreamS2 {
-				t.Errorf("expected incoming edge from %q to stage-2, but not found", tc.wantUpstreamSource)
-			}
-		})
-	}
-}
-
-func TestResolveGraph_FanInCycle_SplitProducerFanInRouting(t *testing.T) {
-	tagA := NewTag[any]("tag-a")
-	tagB := NewTag[any]("tag-b")
-
-	testCases := []struct {
-		name                 string
-		initialTasks         []UntypedTask
-		availableTasks       []UntypedTask
-		wantTaskIDs          []string
-		wantDownstreamSource string
-	}{
-		{
-			name: "downstream task consumes tag provided by split multi-stage consumer",
-			initialTasks: []UntypedTask{
-				createMockTask("prod-bootstrap", "default", nil, ProvidesTag(tagA, WithTagPriority(10))),
-				createMockTask("consumer", "default", []Dependency{
-					tagA.Ref(),
-				}, ProvidesTag(tagB), AllowMultiStageExecution()),
-				createMockTask("prod-feedback", "default", []Dependency{
-					taskid.NewTaskReference[any]("consumer"),
-				}, ProvidesTag(tagA, WithTagPriority(100))),
-				createMockTask("downstream", "default", []Dependency{
-					tagB.Ref(),
-				}),
-			},
-			availableTasks: []UntypedTask{
-				createMockTask("prod-bootstrap", "default", nil, ProvidesTag(tagA, WithTagPriority(10))),
-				createMockTask("consumer", "default", []Dependency{
-					tagA.Ref(),
-				}, ProvidesTag(tagB), AllowMultiStageExecution()),
-				createMockTask("prod-feedback", "default", []Dependency{
-					taskid.NewTaskReference[any]("consumer"),
-				}, ProvidesTag(tagA, WithTagPriority(100))),
-				createMockTask("downstream", "default", []Dependency{
-					tagB.Ref(),
-				}),
-			},
-			wantTaskIDs: []string{
-				"prod-bootstrap#default",
-				"consumer#default-stage-1",
-				"prod-feedback#default",
-				"consumer#default-stage-2",
-				"downstream#default",
-			},
-			wantDownstreamSource: "consumer#default-stage-2",
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			taskSet, err := ResolveGraph(tc.initialTasks, tc.availableTasks, nil)
-			if err != nil {
-				t.Fatalf("unexpected error: %v", err)
-			}
-
-			var gotTaskIDs []string
-			for _, task := range taskSet.GetAll() {
-				gotTaskIDs = append(gotTaskIDs, task.UntypedID().String())
-			}
-
-			if diff := cmp.Diff(tc.wantTaskIDs, gotTaskIDs); diff != "" {
-				t.Errorf("task order mismatch (-want +got):\n%s", diff)
-			}
-
-			incoming := taskSet.IncomingEdges("downstream#default")
-			if len(incoming) != 1 {
-				t.Fatalf("downstream incoming edges count = %d, want 1", len(incoming))
-			}
-			if gotSource := incoming[0].SourceImplID; gotSource != tc.wantDownstreamSource {
-				t.Errorf("downstream incoming edge source = %q, want %q", gotSource, tc.wantDownstreamSource)
-			}
-		})
-	}
-}
