+++ pkg/core/task/graphresolver_stage.go
@@ -15,10 +15,40 @@
 package coretask
 
 import (
+	"context"
 	"slices"
 
+	"github.com/GoogleCloudPlatform/khi/pkg/common/typedmap"
 	"github.com/GoogleCloudPlatform/khi/pkg/core/task/taskid"
 )
+
+// stageTask wraps an UntypedTask to provide a distinct stage implementation ID for multi-stage execution.
+type stageTask struct {
+	originalTask UntypedTask
+	stageID      taskid.UntypedTaskImplementationID
+}
+
+var _ UntypedTask = (*stageTask)(nil)
+
+// UntypedID implements UntypedTask.
+func (s *stageTask) UntypedID() taskid.UntypedTaskImplementationID {
+	return s.stageID
+}
+
+// Labels implements UntypedTask.
+func (s *stageTask) Labels() *typedmap.ReadonlyTypedMap {
+	return s.originalTask.Labels()
+}
+
+// Dependencies implements UntypedTask.
+func (s *stageTask) Dependencies() []Dependency {
+	return s.originalTask.Dependencies()
+}
+
+// UntypedRun implements UntypedTask.
+func (s *stageTask) UntypedRun(ctx context.Context) (any, error) {
+	return s.originalTask.UntypedRun(ctx)
+}
 
 type stageTaskPair struct {
 	stage1 UntypedTask
